requirements.txt
streamlit==0.52.2
wordcloud==1.6.0
pandas==0.25.3
plotly==4.4.1
matplotlib==3.1.2
Pillow==6.2.1





